+++ lib/commercetools-api/src/Models/Inventory/InventoryEntryBuilder.php
@@ -101,6 +101,18 @@ final class InventoryEntryBuilder implements Builder
 
      * @var ?int
      */
+    private $minCartQuantity;
+
+    /**
+
+     * @var ?int
+     */
+    private $maxCartQuantity;
+
+    /**
+
+     * @var ?int
+     */
     private $restockableInDays;
 
     /**
@@ -234,6 +246,28 @@ final class InventoryEntryBuilder implements Builder
     public function getAvailableQuantity()
     {
         return $this->availableQuantity;
+    }
+
+    /**
+     * <p>Minimum quantity that can be added to a Cart. See <a href="/../api/carts-orders-overview#quantity-limits">Quantity limits</a>.</p>
+     *
+
+     * @return null|int
+     */
+    public function getMinCartQuantity()
+    {
+        return $this->minCartQuantity;
+    }
+
+    /**
+     * <p>Maximum quantity that can be added to a Cart. See <a href="/../api/carts-orders-overview#quantity-limits">Quantity limits</a>.</p>
+     *
+
+     * @return null|int
+     */
+    public function getMaxCartQuantity()
+    {
+        return $this->maxCartQuantity;
     }
 
     /**
@@ -391,6 +425,28 @@ final class InventoryEntryBuilder implements Builder
     }
 
     /**
+     * @param ?int $minCartQuantity
+     * @return $this
+     */
+    public function withMinCartQuantity(?int $minCartQuantity)
+    {
+        $this->minCartQuantity = $minCartQuantity;
+
+        return $this;
+    }
+
+    /**
+     * @param ?int $maxCartQuantity
+     * @return $this
+     */
+    public function withMaxCartQuantity(?int $maxCartQuantity)
+    {
+        $this->maxCartQuantity = $maxCartQuantity;
+
+        return $this;
+    }
+
+    /**
      * @param ?int $restockableInDays
      * @return $this
      */
@@ -481,6 +537,8 @@ final class InventoryEntryBuilder implements Builder
             $this->supplyChannel instanceof ChannelReferenceBuilder ? $this->supplyChannel->build() : $this->supplyChannel,
             $this->quantityOnStock,
             $this->availableQuantity,
+            $this->minCartQuantity,
+            $this->maxCartQuantity,
             $this->restockableInDays,
             $this->expectedDelivery,
             $this->custom instanceof CustomFieldsBuilder ? $this->custom->build() : $this->custom